requirements.txt
telethon
python-dotenv
aiofiles